Make all steps of password reset run over SSL
This requires the creation of views with @require_ssl set that calls
into the core django views. Otherwise, when deployed in production,
the middleware will catch these requests and direct them out of SSL.

This has always been a problem, but it broke probably when we turned on
CSRF protection, since the cookie required is no longer passed through.