Maintain redirection URL across oauth signup
Previously when signing up for a new account in the middle of an
authentication process, the final redirection URL was lost and the user
was sent back to the /account/ page. Instead, we now redirect to the
proper URL (for sending back to the community auth client site) if it's
available in the session, both for successful signup and for
canceled logins.