Don't require first and last name fields in OAuth signup
I had fixed the OAuth process to deal with this case, but failed to
cover the actual new account creation part which still required it. With
this fix, we allow a signup without a name. If the user wishes to set a
full name on the account it can be done on the profile page for the
user, but it's possible to proceed to access restricted resources
without doing it.

Reported by @Smylers2 on twitter, relayed by @TokenScandi