Reference OLD instead of NEW in trigger
The trigger is used for both UPDATE and DELETE, but in DELETE there is
no value in OLD.

Given that it only references the user_id field and this field cannot
change, we can just use OLD in both the UPDATE and DELETE case.

Back-patching in existing migration since it hasn't really been deployed
anywhere yet.