Nicer error message when consent form is messed with
Instead of an exception complaining about bad style URLs, just ensure
that the URL for the next parameter is always relative. (The form for
consent can only be triggered via one redirect, and it always has the
parameter relative).

We did the right thing before (as in, did not perform a redirect), but
the error dump was not nice.