Bugfix: Correctly sort vote options by preference
Vote options, when pulled from the database, were sorted by id and not
preference. While they were on the vast majority stored by preference
order in the database anyway, this did not matter most of the
time. However, in rare occasions, this was not the case. For instance,
this changed the vote of a single voter (!) in the 2007 board election
(their options were sorted by alphabetical order of preference,
i.e. 10 11 12 13 1 2 3 4 5 6 7 8 9), affecting the displayed result of
the election.